Reorganize file structure and replace old user profile with new
